수정이 완료된 내용은 아래와 같으며, CreateManager 수정을 진행 중입니다.
 * ItemData_Weapon.cs v2.0 - 각인배열 인덱스 오류 수정, 각인 최대수량 프로퍼티 추가, 속성석 적용 메서드 추가
 * ItemData_Misc.cs v3.0 - 속성석입력 받을 때 이름에 따른 분기로 변경, 각인석 정보 구조체에서 상태창 이미지 인덱스 정보를 가지도록 변경
 * ItemInfo V8.1 - UpdateInfo에서 상태창 이미지까지 반영,
 * ItemPointerStatusWindow v3.0 - statusWindow, 각인석 이미지를 iicMiscOther에서 구조체 인덱스로 접근하여 반영
 * InventoryManagement.cs v4.0 - 슬롯리스트가 개별탭 별로 있던 것을 전체탭만 존재하도록 축소 및 개별 탭의 갯수만큼만 슬롯이 보이도록 변경
 * Drop.cs v3.1 - 속성석 5개 분기처리, 싱글톤 참조 삭제, 직접 BasePerformance 조정 삭제, -1감소 리스트에서 딕셔너리로 수정
 * GameManager_part2.cs v1.0 - 게임상태 변수 isNewGame과 버튼 시작, 이어하기 기능에서 isNewGame을 수정하도록 구현

+++ Assets/02.Scripts/DataInfo/GameData.cs
@@ -24,6 +24,10 @@ using CraftData;
  * 2- SerializableInventory 클래스를 Inventory 파일로 위치로 옮김.
  * 씬 전환시에도 사용하기 위해
  * 3- 플레이어 위치, 회전 정보를 담기 위한 STransform 클래스 정의
+ * 
+ * <v2.1 - 2023_1218_최원준>
+ * 1- 금화 은화 변수 int형으로 변경
+ * 
  */
 
 namespace DataManagement
@@ -52,12 +56,12 @@ namespace DataManagement
         /// <summary>
         /// 금화
         /// </summary>
-        public float gold;
+        public int gold;
 
         /// <summary>
         /// 은화
         /// </summary>
-        public float silver;
+        public int silver;
 
         /// <summary>
         /// 플레이어의 제작 가능 목록과 숙련도, 레시피 맞춘 횟수 정보가 담겨 있는 제작 관련 클래스 입니다.
